refactor route classes server
